Moves blocks views and tests into preferences dir
